Updates to benchmark scripts
